refactor: 베팅 금액 타입 long 타입으로 변경
- 30억이 넘는 금액이 베티 금액으로 들어오면 int 타입을 사용했을때 범위를 초과하는 상황이 발생함
- long 타입을 사용하여 더 큰 범위의 베팅 금액을 받을 수 있음

+++ src/main/java/controller/BlackJackController.java
@@ -48,13 +48,13 @@ public class BlackJackController {
     }
 
     private void profitResult(final BetSystem betSystem, final Dealer dealer, final List<Player> players) {
-        Map<Gamer, Integer> gamerIntegerMap = betSystem.calculateProfit(dealer, players);
+        Map<Gamer, Long> gamerIntegerMap = betSystem.calculateProfit(dealer, players);
         outputView.printProfitResult(gamerIntegerMap);
     }
 
     private boolean isPrepareBlackjack(final Dealer dealer, final List<Player> players, final BetSystem betSystem) {
         if (isPrepareCardsBlackjack(dealer, players)) {
-            Map<Gamer, Integer> gamerIntegerMap = betSystem.calculateProfit(dealer, players);
+            Map<Gamer, Long> gamerIntegerMap = betSystem.calculateProfit(dealer, players);
             outputView.printCardResult(dealer, players);
             outputView.printProfitResult(gamerIntegerMap);
             return true;
